refactor(backend): moved schemas to separate files
