+++ new/src/backend/app/db/alembic/env.py
@@ -10,12 +10,9 @@ from alembic import context
 
 from db.schemas import Base  # noqa: F401
 
-# from db.schemas.calculation.calculation import Calculation  # noqa: F401
-# from db.schemas.calculation.calculation_config import CalculationConfig  # noqa: F401
-# from db.schemas.calculation.calculation_set import CalculationSet  # noqa: F401
-# from db.schemas.user.user import User  # noqa: F401
-# from db.schemas.moleculeset_stats import MoleculeSetStats, AtomTypeCount  # noqa: F401
-from db.schemas.schema import *  # noqa: F401
+from db.schemas.calculation import *  # noqa: F401
+from db.schemas.stats import *  # noqa: F401
+from db.schemas.user import *  # noqa: F401
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
