+++ new/src/backend/app/db/repositories/user_repository.py
@@ -1,4 +1,4 @@
-from db.schemas.schema import User
+from db.schemas.user import User
 from db.database import SessionManager
 
 from sqlalchemy import select
