+++ new/src/backend/app/db/schemas/stats.py
@@ -1,7 +1,7 @@
 import uuid
 
 import sqlalchemy as sa
-from sqlalchemy.orm import Mapped, mapped_column, relationship
+from sqlalchemy.orm import Mapped, relationship, mapped_column
 
 from db.schemas import Base
 
@@ -15,15 +15,13 @@ class AtomTypeCount(Base):
     symbol: Mapped[str] = mapped_column(sa.VARCHAR(10), primary_key=False)
     count: Mapped[int] = mapped_column(sa.Integer, nullable=False)
 
-    molecule_set_id = mapped_column(sa.VARCHAR(100), sa.ForeignKey("molecule_set_stats.file_hash"))
+    molecule_set_id = mapped_column(
+        sa.VARCHAR(100), sa.ForeignKey("molecule_set_stats.file_hash"), nullable=False
+    )
     molecule_set_stats = relationship("MoleculeSetStats", back_populates="atom_type_counts")
 
     def __repr__(self):
-        return f"""<AtomTypeCount
-            id={self.id},
-            symbol={self.symbol},
-            count={self.count},
-            molecule_set_id={self.molecule_set_id}"""
+        return f"""<AtomTypeCount id={self.id}, symbol={self.symbol}, count={self.count}, molecule_set_id={self.molecule_set_id}"""
 
 
 class MoleculeSetStats(Base):
@@ -40,10 +38,9 @@ class MoleculeSetStats(Base):
     total_atoms: Mapped[int] = mapped_column(sa.Integer, nullable=False)
 
     atom_type_counts = relationship("AtomTypeCount", back_populates="molecule_set_stats")
+    calculation_sets = relationship(
+        "CalculationSet", secondary="calculation_set_stats", back_populates="molecule_set_stats"
+    )
 
     def __repr__(self):
-        return f"""<MoleculeSetStats
-            file_hash={self.file_hash},
-            total_molecules={self.total_molecules},
-            total_atoms={self.total_atoms},
-            atom_type_counts={self.atom_type_counts}"""
+        return f"""<MoleculeSetStats file_hash={self.file_hash}, total_molecules={self.total_molecules}, total_atoms={self.total_atoms}, atom_type_counts={self.atom_type_counts}"""
